# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Цифровая станция — H18",
    "description": "Physical digital twin, independent validator, isolated deterministic heuristic planner and SSE.",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "summary": "Login",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/snapshot": {
      "get": {
        "summary": "Snapshot",
        "operationId": "snapshot_api_v1_snapshot_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/State"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/replans": {
      "post": {
        "summary": "Replan",
        "operationId": "replan_api_v1_replans_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplanInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandReceiptResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/incidents": {
      "post": {
        "summary": "Create Incidents",
        "operationId": "create_incidents_api_v1_incidents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentBatchInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandReceiptResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/replans/{job_id}": {
      "get": {
        "summary": "Replan Detail",
        "operationId": "replan_detail_api_v1_replans__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplanDetail"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "APIError": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorPayload"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "APIError"
      },
      "AuthResponse": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "user"
        ],
        "title": "AuthResponse"
      },
...
      "CommandReceiptResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "input_revision": {
            "type": "integer",
            "title": "Input Revision"
          },
          "state_version": {
            "type": "integer",
            "title": "State Version"
          },
          "result": {
            "additionalProperties": true,
            "type": "object",
            "title": "Result"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "run_id",
          "input_revision",
          "state_version",
          "result"
        ],
        "title": "CommandReceiptResponse"
      },
...
      "ErrorPayload": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "details",
          "request_id"
        ],
        "title": "ErrorPayload"
      },
...
      "IncidentBatchInput": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "expected_input_revision": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Expected Input Revision"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/IncidentInput"
            },
            "type": "array",
            "maxItems": 10,
            "minItems": 1,
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "run_id",
          "expected_input_revision",
          "items"
        ],
        "title": "IncidentBatchInput"
      },
...
      "LoginInput": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "format": "password",
            "title": "Password",
            "writeOnly": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginInput"
      },
...
      "ReplanDetail": {
        "properties": {
          "job": {
            "$ref": "#/components/schemas/ReplanJob"
          },
          "plans": {
            "items": {
              "$ref": "#/components/schemas/PlanDetail"
            },
            "type": "array",
            "title": "Plans"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "job",
          "plans"
        ],
        "title": "ReplanDetail"
      },
      "ReplanInput": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "expected_input_revision": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Expected Input Revision"
          },
          "reason": {
            "type": "string",
            "const": "manual",
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "run_id",
          "expected_input_revision",
          "reason"
        ],
        "title": "ReplanInput"
      },
...
      "State": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "scenario_id": {
            "type": "string",
            "title": "Scenario Id"
          },
          "scenario_epoch": {
            "type": "string",
            "title": "Scenario Epoch"
          },
          "event_seq": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Event Seq"
          },
          "state_version": {
            "type": "integer",
            "minimum": 0.0,
            "title": "State Version"
          },
          "input_revision": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Input Revision"
          },
          "config_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Config Version"
          },
          "server_time": {
            "type": "string",
            "title": "Server Time"
          },
          "sim_time_s": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Sim Time S"
          },
          "speed": {
            "type": "integer",
            "enum": [
              1,
              5,
              10
            ],
            "title": "Speed"
          },
          "mode": {
            "type": "string",
            "enum": [
              "paused",
              "running"
            ],
            "title": "Mode"
          },
          "station": {
            "$ref": "#/components/schemas/Station"
          },
          "active_plan_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Plan Id"
          },
          "tracks": {
            "items": {
              "$ref": "#/components/schemas/Track"
            },
            "type": "array",
            "title": "Tracks"
          },
          "zones": {
            "items": {
              "$ref": "#/components/schemas/Zone"
            },
            "type": "array",
            "title": "Zones"
          },
          "trains": {
            "items": {
              "$ref": "#/components/schemas/Train"
            },
            "type": "array",
            "title": "Trains"
          },
          "wagon_groups": {
            "items": {
              "$ref": "#/components/schemas/WagonGroup"
            },
            "type": "array",
            "title": "Wagon Groups"
          },
          "resources": {
            "items": {
              "$ref": "#/components/schemas/Resource"
            },
            "type": "array",
            "title": "Resources"
          },
          "operations": {
            "items": {
              "$ref": "#/components/schemas/Operation"
            },
            "type": "array",
            "title": "Operations"
          },
          "incidents": {
            "items": {
              "$ref": "#/components/schemas/Incident"
            },
            "type": "array",
            "title": "Incidents"
          },
          "conflicts": {
            "items": {
              "$ref": "#/components/schemas/Conflict"
            },
            "type": "array",
            "title": "Conflicts"
          },
          "plans": {
            "items": {
              "$ref": "#/components/schemas/PlanSummary"
            },
            "type": "array",
            "title": "Plans"
          },
          "last_replan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReplanJob"
              },
              {
                "type": "null"
              }
            ]
          },
          "efficiency": {
            "$ref": "#/components/schemas/Efficiency"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "schema_version",
          "run_id",
          "scenario_id",
          "scenario_epoch",
          "event_seq",
          "state_version",
          "input_revision",
          "config_version",
          "server_time",
          "sim_time_s",
          "speed",
          "mode",
          "station",
          "active_plan_id",
          "tracks",
          "zones",
          "trains",
          "wagon_groups",
          "resources",
          "operations",
          "incidents",
          "conflicts",
          "plans",
          "last_replan",
          "efficiency"
        ],
        "title": "State"
      },
      "Station": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "parks": {
            "items": {
              "$ref": "#/components/schemas/Park"
            },
            "type": "array",
            "title": "Parks"
          },
          "layout": {
            "$ref": "#/components/schemas/Layout"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "parks",
          "layout"
        ],
        "title": "Station"
      },
      "Track": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "park_id": {
            "type": "string",
            "title": "Park Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "receiving_departure",
              "lead",
              "sorting",
              "cargo",
              "parking"
            ],
            "title": "Kind"
          },
          "length_m": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Length M"
          },
          "availability": {
            "type": "string",
            "enum": [
              "open",
              "closed",
              "closure_pending"
            ],
            "title": "Availability"
          },
          "occupied_length_m": {
            "type": "number",
            "minimum": 0.0,
            "title": "Occupied Length M"
          },
          "train_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Train Ids"
          },
          "group_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Group Ids"
          },
          "locomotive_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Locomotive Ids"
          },
          "assigned_train_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Train Id"
          },
          "active_operation_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Active Operation Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "park_id",
          "kind",
          "length_m",
          "availability",
          "occupied_length_m",
          "train_ids",
          "group_ids",
          "locomotive_ids",
          "assigned_train_id",
          "active_operation_ids"
        ],
        "title": "Track"
      },
...
      "User": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "role": {
            "type": "string",
            "enum": [
              "viewer",
              "operator",
              "dispatcher",
              "admin"
            ],
            "title": "Role"
          },
          "resource_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Resource Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "display_name",
          "role",
          "resource_ids"
        ],
        "title": "User"
      },
...
    }
  }
}